Update
Changed the naming of argument 0 for the Alert method to account for non-strings.
Reduced font size by 2.
Reduced timer by 1 second.
Updated description for alert method.
Added a check if the string_or_format is an empty string, and updated the generated alert message.
Changed the vertical text alignment from middle to top.